Disable enterprise tests in CI but have a way of running them in dev
