+++ README.md
@@ -4,7 +4,7 @@
 
 # Requirements
 
-- Go > 1.20
+- Go > 1.18
 
 # OpenAPI generator used
 
@@ -41,7 +41,9 @@ go run main.go -unleash-api http://localhost:4242 -authorization *:*.unleash-ins
 
 # Testing
 
-Testing relies on a running environment. You can use docker-compose.yml file to spin up an environment with a clean state and predictable admin API token
+Testing relies on a running environment. You can use docker-compose.yml file to spin up an environment with a clean state and predictable admin API token.
+
+**Note**: some tests rely on an enterprise version of Unleash. To run those tests locally you need to set the environment variable `UNLEASH_ENTERPRISE=true`.
 
 A basic test is automatically generated by the tool (find them under `client/test`), but they don't have a default configuration and doesn't have payloads required for some of the endpoints. This is the reason why the tools marks all of them as Skip. Still they're a good starting point.
